Fix a case were sorting auras would result in spacing.
This was because the sort function potentially could move icons outside the range :SetAuraPosition() would operate on. A simple example of this happening would be:
Pre sorting:
1:'visibleAura'
2:'visibleAura'
3:'inactiveAura'
4:'inactiveAura'
5:'inactiveAura'

Post sorting:
1:'visibleAura'
2:'inactiveAura'
3:'inactiveAura'
4:'visibleAura'
5:'inactiveAura'

The iterator would only try to position icon 1 and 2, thus leaving a space of two empty icons. The position remained correct due to the fact that oUF doesn't clear old positions. Ironically the fix is a ~3 liner...