Clean up the file.
The entire file has been converted from dos LE to unix, which is consistent with the rest of oUF. The file has also been re-indented with tabs instead of spaces. Excessive use of spaces at end of lines have been removed. A round of sed has also removed some of the syntax differences so it's more consistent with the coding style used in oUF.