portrait: Reset settings before setting the new model.
Reseting the camdistancescale vales etc is needed before reseting the model otherwise it will not reload properly if you change from a gnome to a dwarf etc.

I originally thought that clearing the model would wipe the model attributes but that is not the case. The attributes have to be set before the model rendering.

http://www.wowinterface.com/forums/showpost.php?p=246077&postcount=17