tags: Remove 1/2/3 static cases
We'll be losing ~18% performance here for the sake of code readability.
While that might sound scary, in real world performance that's only
0.26ms over 360 calls, which is the worst realistic scenario. That's 90
unit frames with 4 tagged font strings each of which has 3 tag
functions. That's not enough to lose even 1 full frame at 60 fps, but at
that point your fps will tank due to all the mobs running on the screen
anyway.