power: Fix colouring by power type logic
First of all, it never did what it's supposed to do since pToken was
never nil. Secondly, when UnitPowerType returns info for a custom colour
it hijacks power type 3, which normally is "ENERGY", but when it's
overridden it becomes "POWER_TYPE_*", for instance,
"POWER_TYPE_FEL_AMMO" and "POWER_TYPE_POWER" for the 1st and 2nd HFC
encounters respectively, and the method also provides alternative
r, g, and b values. Technically, this should be enough to cover
everything.